Starting refactoring for better testability of webviews logic
